+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0
+
+- Upgrades iOS SDK to https://github.com/RevenueCat/purchases-ios/releases/tag/2.5.0
+- Upgrades Android SDK to https://github.com/RevenueCat/purchases-android/releases/tag/2.3.0.
+- Adds Facebook as supported attribution network.
+- Adds automatic Apple Search Ads attribution collection. Disabled by default.
+- Adds introductory pricing to the iOS product.
+
 ## 0.1.4
 
 - Fixes UninitializedPropertyAccessException when trying to close the instance onDestroy of the Android Activity.